requirements.txt
langchain
langchain-community
langchain-openai
fireworks-ai
python-dotenv
pyyaml
hydra-core
omegaconf
pydantic
duckduckgo-search
chromadb
langchain-chroma
langchain-experimental
pdfplumber